Fixed selection tools not being able to select highest row or column of pixel

+++ src/Tools/FillSelection.hpp
@@ -20,13 +20,13 @@ namespace Mimp {
 		//! @param toolBox ToolBox containing the FillSelection Tool
 		FillSelection(ToolBox &);
 
-		void onMouseDrag(Vector2<int>, Vector2<int>, MouseClick, Image &) override {};
+		void onMouseDrag(Vector2<float>, Vector2<float>, MouseClick, Image &) override {};
 
 		//! @brief Handle the mouse click of the Tool.
 		//! @param pos Position of the tool
 		//! @param click Mouse click state
 		//! @param image The Image to edit
-		void onClick(Vector2<int> pos, MouseClick click, Image &image) override;
+		void onClick(Vector2<float> pos, MouseClick click, Image &image) override;
 
 		//! @brief Get the parameters panel for the Tool.
 		//! @return tgui::ScrollablePanel::Ptr Pointer containing the parameters panel
